feat(gateway): Add service for deleting users
- Add mock function and real function for user updating service
- Create unified service function based on environment
