+++ gateway/src/graphql/resolvers/user/services/index.ts
@@ -1,4 +1,5 @@
 import createUser from "./create-user";
 import updateUser from "./update-user";
+import deleteUser from "./delete-user";
 
-export default { createUser, updateUser };
+export default { createUser, updateUser, deleteUser };
